+++ src/libs/piano/Piano/index.ts
@@ -1,0 +1,1 @@
+export { Piano } from './Piano';
