+++ src/libs/piano/PianoKey/index.ts
@@ -1,0 +1,2 @@
+export { BlackPianoKey } from './BlackPianoKey';
+export { WhitePianoKey } from './WhitePianoKey';
