+++ src/libs/piano/PianoKeyGroup/index.ts
@@ -1,0 +1,1 @@
+export { PianoKeyGroup } from './PianoKeyGroup';
