+++ src/libs/piano/PianoKeySeperator/index.ts
@@ -1,0 +1,1 @@
+export { PianoKeySeperator } from './PianoKeySeperator';
